Add files via upload
This contains the raw .NC files for SSP" and SSP5 precipitation models above Norway. Specific locations can be extracted to Excel using the NETCDF_Extract.py script. SSP2Annual and SSP5 Annual contain the extracted results, which come in the form of kg.m2.s-1; these have been converted to mm, then divided by 12 to receive precipitation as the monthly average per year, labelled under the column 'precipitation'. This column is what should be referred to when running any of the models that include precipitation.