CMIP future precipitation data
This precipitation data is for the use of future modelling. The raw SSP2 and SSP5 data for Norway are found in the .NC files, these can be reduced to the specific area of Engabreen and the power plants using the 'NETCDF_EXTRACT.py' script. The 'SSP2Annual' and 'SSP5Annual' files contain precipitation data for each region in kg/m²/s-1, which is then converted to mm and divided by 12 to obtain the monthly average per year. This value is stored under the 'precipitation' heading and is what should be used in the model.