Historical data for power plants and Engabreen
These files contain the historical energy production at Glomfjord and Svartisen power plants, along with precipitation in the region. Historical annual mass balance, length change, and summer mass balance data are also included in the respective named files. This data is intended to be used in the model to identify relationships and validate the model's performance. The code uploaded for each model is written with the intention of use with annual mass balance, so adjustments to which variables to use need to be implemented in the appropriate sections of the model's code.